1. fix

+++ clickhouse_mysql/tablesqlbuilder.py
@@ -268,7 +268,7 @@ ORDER BY id
         elif  mysql_type.startswith('DEC') or mysql_type.startswith('FIXED') or mysql_type.startswith('NUMERIC'):
             ch_type = 'String'
         elif mysql_type.startswith('FLOAT'):
-            ch_type = 'Float32'
+            ch_type = 'Float64'
         elif mysql_type.startswith('DECIMAL') or mysql_type.startswith('DOUBLE') or mysql_type.startswith('REAL'):
             ch_type = 'Float64'
